+++ src/qg/apps/bfabric_app_editor.py
@@ -3,6 +3,9 @@
 from pathlib import Path
 
 from qg.apps._bfabric_auth import create_bfabric_fastapi_app
+from qg.logging_setup import configure_logging
+
+configure_logging()
 
 app = create_bfabric_fastapi_app(
     Path(__file__).parent / "config_editor.py",